+++ release-notes.md
@@ -1,3 +1,12 @@
+# 2.0.0 Beta 19 - Mar 12th 2013
+
+- Fix stylus execution outside of the lumbar directory
+- Add overrides false flag to prevent output of particular resources from mixins
+
+# 2.0.0 Beta 17 - Feb 21st 2013
+
+- Added `lumbar` grunt task
+
 # 2.0.0 Beta 15 - Feb 5th 2013
 
 Changes: